+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Automatic carrier validation and color assignment after network loading
+
 ## [0.1.0] - 2024-10-04
 
 ### Added